Optimized Profile & Dashboard UI

+++ backend/client/src/actions/profile.js
@@ -8,15 +8,12 @@ import {
   UPDATE_PROFILE,
   CLEAR_PROFILE,
   GET_REPOS,
-  ACCOUNT_DELETED,
-  LOADING
+  ACCOUNT_DELETED
 } from './types'
 
 // Get current user profile
 export const getCurrentProfile = () => async dispatch => {
   try {
-    dispatch({ type: CLEAR_PROFILE });
-    dispatch({ type: LOADING });
     const res = await axios.get('/api/profile/me');
     dispatch({ type: GET_PROFILE, payload: res.data });
   } catch (e) {
@@ -139,27 +136,31 @@ export const addEducation = (formData, history) => async dispatch => {
 
 // Delete Experience
 export const deleteExperience = id => async dispatch => {
-  try {
-    const res = await axios.delete(`/api/profile/experience/${id}`);
+  if(window.confirm('Are you sure? This action cannot be undone.')) {
+    try {
+      const res = await axios.delete(`/api/profile/experience/${id}`);
 
-    dispatch({ type: UPDATE_PROFILE, payload: res.data });
+      dispatch({ type: UPDATE_PROFILE, payload: res.data });
 
-    dispatch(setAlert('Experience Deleted!', 'success', 2000));
-  } catch (e) {
-    dispatch({ type: PROFILE_ERROR, payload: {msg: e.response.statusText, status: e.response.status} });
+      dispatch(setAlert('Experience Deleted!', 'success', 2000));
+    } catch (e) {
+      dispatch({ type: PROFILE_ERROR, payload: {msg: e.response.statusText, status: e.response.status} });
+    }
   }
 }
 
 // Delete Education
 export const deleteEducation = id => async dispatch => {
-  try {
-    const res = await axios.delete(`/api/profile/education/${id}`);
+  if(window.confirm('Are you sure? This action cannot be undone.')) {
+    try {
+      const res = await axios.delete(`/api/profile/education/${id}`);
 
-    dispatch({ type: UPDATE_PROFILE, payload: res.data });
+      dispatch({ type: UPDATE_PROFILE, payload: res.data });
 
-    dispatch(setAlert('Education Deleted!', 'success', 2000));
-  } catch (e) {
-    dispatch({ type: PROFILE_ERROR, payload: {msg: e.response.statusText, status: e.response.status} });
+      dispatch(setAlert('Education Deleted!', 'success', 2000));
+    } catch (e) {
+      dispatch({ type: PROFILE_ERROR, payload: {msg: e.response.statusText, status: e.response.status} });
+    }
   }
 }
